Remove TempestRegexp argument
Looks like it was not in use since the begning.
Users usually finds simpler to have explicit negative list,
even tough it is possible to consutuct a singe regexp.
Also possible to construct a single regexp on the operator side,
in case other way of passing would be prefered later on.

Moving the default allowedTests to the defaults.

Note: Consider moving the tempest run arguments
to sub structure.