Add tempest options
This patch introduces a set of parameters for the tempestRun section
in the Tempest CR. This field parameters allow a user to specify
parameters with which tempest command should be executed inside
the Tempest pod.

The documentation is updated as well so that it demonstrates how the
new field should be utilized.