Change link to ci-framework tempest role
In https://github.com/openstack-k8s-operators/ci-framework/pull/943
the roles were moved from the ci_framework folder to the repo root to
follow the structure of an ansible collection. Since eventually the
ci_framework folder will be delete, this change points to the new role
location to avoid a dead link in the future.