Update RBAC rules for the test-operator
This patch updates the RBAC rules for the test-operator so that
the controller can list persistent volume claims in the environment.

The change is needed in order to be able to sucessfully create
pvcs from withing the operator. This patch is part of sequence of
patches that test update the rbac rules.