Make tobiko part of the test-operator
This patch introduces tobiko to the test-operator. The patch adds a new
CRD called Tobiko and introduces a new controller that manages it. The
user can use the new CRD to spawn tobiko tests.

To ensure that there is always one framework running inside the cluster
a locking mechanism based on config-maps is introduced. This locking
mechanism is based on ensuring that there is always only one cm named
test-operator-lock in the cluster. The controller which created the cm can
execute the tests.