[Tobiko] Add support for workflows
This patch introduces workflows to the Tobiko CR. The workflows allow
to specify multiple executions of tobiko in a single CR. The steps
for a workflow are specified in a newly introduced Workflow section.
Values specified in the Workflow section have a higher priority and
overwrite configuration in the Spec section.

Depends-On: https://github.com/openstack-k8s-operators/tcib/pull/146