[Tempest] Install extra RPMs
This patch introduces tempestRun.extraRPMs parameter in the Tempest CR
that allows to specifying a list of extra RPMs that should be installed
before the execution of Tempest.

This parameter accepts a list of URLs that point to the RPMs that
should be installed. The RPMs are installed in the order in which they
are listed via the parameter.

Depends-On: https://github.com/openstack-k8s-operators/tcib/pull/175