Add --create when tempestConf section is empty
When tempestConf section is empty in Tempest CR the default value
for create is set to false, which leads to discover-tempest-config
failure, even though the CR is valid. This patch sets create to
true if the tempestConf section is empty to prevent the failure.

Additionally, this patch corrects local configuration of webhooks
as this patch works with mutating webhook and needs these changes
to work properly.