Request separate PVCs when parallel is set to true
This patch changes the way PVCs are requested when the parallel
option is set to true. Instead of using one PV for all the workflow
steps, it is divided into multiple PVs. This change is needed,
because the LVM backed storage does not support the RWX, which
causes issues in some cases.