Consume test-operator-config produced by openstack-operator
This patch in openstack-operator [1] introduces test-operator-config
CM. This CM among other things contains information about which
container image should be used for testing.

This patch ensures that we consume this information. The image
that is used for testing is picked up from these 3 places (sorted
from highest to lowest priority).

 1. containerImage value in test-operator related CR
   (instance.Spec.ContainerImage)
 2. value from the test-operator-config
 3. value from RELATED_IMAGE_ env variables

[1] https://github.com/openstack-k8s-operators/openstack-operator/pull/823