Disable AllowPrivilegeEscalation
This PR ensures that test pods do not run with allowPrivilegeEscalation.
In order to be able to run pods without the privilege escalation flag,
the following changes had to be made:

- Removal of mounting of the localtime and machine-id files (required
  hostPath). These files are not required for successfull functioning
  of the test-operator.

- Removal of the CAP_AUDIT_WRITE flag.

- Dropping of of all capabilities by default.

However, for certain set of test-operator functionalities the parameters
mentioned above are required to be enabled. Because of that we are
introducing insecure flag that can be turnd on whenever functionality
,for example like extraRPMs, is required.