Skip verify_metadata_from_api
This is a new test let's skip this for now until the failure of
the test is resolved. It seems like the failure is not related
to the test-operator as the test is capable of connecting to the
VM through ssh.

We can unskip it once the issue is resolved.