Increase ResourceMemory limit
We are hitting an issue when pods are being killed in jobs because of
OOM error. It seems like the ResourceMemory limit set with this PR [1]
is not enough for the test operator pods.

This patch increases the ResourceMemory limit to 8Gi. We should revisit
this problem later and investigate whether the limit can not be lower and
what exactly caused the OOM errors.

[1] https://github.com/openstack-k8s-operators/test-operator/pull/205