Modify SELinuxLevel to SELinuxOptions
This commit removest the SELinuxLevel and replaces it with
SELinuxOptions. This new parameter is added to all test-operator
CRs.

The motivation behind this change is that we want to get rid of
the default value for the SELinuxLevel. SELinuxLevel should be
set only when privileged: true (meaning the scc anyuid/privileged
is used). Also, move to SELinuxOptions exposes all SELinux
to the user.