Change default value of SELinuxLevel
This commit removes the default value of SELinuxLevel. This option
should be only used when privileged: true because scc privileged
is required when setting SELinuxLevel on a pod.

Setting SELinuxLevel is sometimes required in order to be able to
allow RWX for a mounted PVC. We're going to move the setting of the
SELinuxLevel to ci-framework and leave the default parameters for
the test-operator secure.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/2448