Fix the nil pointing issue
When a workflow is set without ExtraConfigMap in ansible controller the
pod won't be able to start because the pointer won't be pointing at
nothing. Changed the condition to allow avoiding this issue.

not tested, waiting for a test enviornment to work on this

Jira: https://issues.redhat.com/browse/OSPRH-11191