Run test pods with readOnlyRootFileSystem: true
Up until now the test pods were being spawned with writable root
file system (readOnlyRootFilesystem: false). This is against the
best security practices as setting the readOnlyRootFileSystem to
false increases the size of the attack surface.

This patch ensures that each pod spawned by the test-operator has
the readOnlyRootFilesystem set to true by default. It is possible
to run a pod with writable roo file system by setting
privileged: true.

Depends-On: https://github.com/openstack-k8s-operators/tcib/pull/233