Remove unnecessary rights
In this PR [1] we dropped the usage of the Jobs by the test-operator.
This allows us to drop the rights for:

 - ServiceAccount managing & creation
 - Role managing & creation
 - RoleBinding managing & creation

These rights were only needed because Jobs were required to be spawned
with an extra ServiceAccount that would have elevated privileges in case
the test pods need to run with privileged SecurityContext.

[1] https://github.com/openstack-k8s-operators/test-operator/pull/266

Depends-On:  https://github.com/openstack-k8s-operators/test-operator/pull/266