Update documentation about local testing
Now that test-operator is installed with other operators, local
testing can't be done without some measures. We do not want our
local test-operator and the test-operator installed by default
to fight for control, so it is needed to disable the default one.
This patch documents how this can be achieved.