Fix behavior for workflows
In current case, when there is a global value specified
on top level (i.e. non-workflow level, that should be applied
to all workflows) and no workflow-specific value is provided,
that global value will always be overridden with the default
from workflow-level (i.e. the global value will be ineffective
when workflows are used). This commit adjusts it, so value
on workflow-level is optional and has no default – in such case
for the workflow-level the value from top level will be applied
(unless something is explicitly defined on workflow-level).