Add warning for ExtraConfigmapsMounts deprecation
Now that test-operator supports ExtraMounts parameter, there is no
need for a parameter adding extra configmap mounts. It can be
directly done through ExtraMounts, so I think of it as a duplicate
code. But it is not a good practice to remove parameters right
away, so I am first adding warning for users who might be using
extraConfigmapsMounts parameters about its deprecation. Let's
remove it after this warning is old enough.