Remove test-operator periodic job
The test-operator periodic job has been failing for a long time.
It was supposed to be used for monitoring, but no one has been
checking it. There could be a fix in form of a workaround. However,
that would add a more complex logic for a feature that will most
likely not be used.